Refactor semantic expression helpers
